+++ developer/15.0/reference/file-types/index.php
@@ -1,0 +1,50 @@
+<?php
+  require_once('includes/template.php');
+
+  head([
+    'title' => "File Types"
+  ]);
+?>
+
+<h1>File Types</h1>
+
+<dl>
+  <dt><a href="ico">ICO files</a></dt>
+  <dd>Keyboard icon files (binary file format)</dd>
+
+  <dt><a href="kmn">KMN files</a></dt>
+  <dd>Keyboard source files (text file format)</dd>
+  
+  <dt><a href="kmp">KMP files</a></dt>
+  <dd>Keyboard package files (zip file format)</dd>
+  
+  <dt><a href="kmx">KMX files</a></dt>
+  <dd>Keyboard files (binary file format)</dd>
+  
+  <dt><a href="kpj">KPJ files</a></dt>
+  <dd>Keyman Developer Project files (XML file format)</dd>
+  
+  <dt><a href="kps">KPS files</a></dt>
+  <dd>Keyboard package source files (XML file format)</dd>
+  
+  <dt><a href="kvk">KVK files</a></dt>
+  <dd>Visual keyboard files (binary file format)</dd>
+
+  <dt><a href="kvks">KVKS files</a></dt>
+  <dd>Visual keyboard source files (XML file format)</dd>
+
+  <dt><a href="model-js">MODEL.JS files</a></dt>
+  <dd>Lexical model files (text file format)</dd>
+
+  <dt><a href="model-ts">MODEL.TS files</a></dt>
+  <dd>Lexical model definition files (text file format)</dd>
+
+  <dt><a href="tsv">TSV files</a></dt>
+  <dd>Word list source files (tab-separated values file format)</dd>
+
+  <dt><a href="ttf">TTF files</a></dt>
+  <dd>TrueType font files (binary file format)</dd>
+
+  <dt><a href="keyman-touch-layout">keyman-touch-layout files</a></dt>
+  <dd>Keyman touch layout source files (JSON file format)</dd>
+</dl>