+++ keyboard/fv_dane_zaa_zaage/9.2/fv_dane_zaa_zaage.php
@@ -1,0 +1,47 @@
+<?php 
+  $pagename = 'Dane-Z̲aa Z̲áágéʔ Keyboard Help';
+  $pagetitle = 'Dane-Z̲aa Z̲áágéʔ Keyboard Help';
+  $pagestyle = <<<END
+  
+#wikitext table { margin: 12px; }
+#wikitext table td { border: solid 1px #888888 !important }
+#wikitext table th { border: solid 1px #888888 !important; background: #ffcccc }
+.lang2 {font-family:"Aboriginal Sans"}  
+END;
+
+  require_once('header.php');
+?>
+
+<p>This keyboard is designed for the <b>Dane-Z̲aa Z̲áágéʔ</b> language of the Western Subarctic region of Canada.</p>
+
+<h2>Desktop Layout</h2>
+<div id='osk' data-states='default shift rightalt rightalt-shift'>
+
+<p>Combining accents are located on the standard <strong>/ and \</strong> keys. Type an accent after a vowel to add it on the vowel. If you've typed the wrong accent, just type the correct accent to replace it.</p>
+<ul>
+  <li>á = a then /</li>
+  <li>ǎ = a then \</li>
+  <li>â = a then Shift-\</li>
+</ul>
+
+<p>By holding the ALT key, you can return to the standard US English keyboard value of any modified key (such as <strong>p, r, q, v, x</strong>, and various punctuation). Shift-Alt will give you the standard English shifted value of the modified key (such as <strong>P, R, Q, V, Z</strong>, etc).</p>
+
+<h2>Mobile/Tablet Touch Layout</h2>
+<div id='osk-tablet' data-states='default shift numeric'>
+
+<p>Accent marks are located in the top row.</p>
+
+<p>The combining ogonek/hook is located in the second row (in place of <strong>f</strong>). The combining low line is located in the third row (in place of <strong>x</strong>).</p>
+
+<p>Combined characters using the ogonek/hook and low line are also available in a long press menu. Keys with long-press options are:</p>
+<ul>
+  <li>a > ą (Ą)</li>
+  <li>d > d̲z̲ (D̲z̲ D̲Z̲)</li>
+  <li>e > ę (Ę)</li>
+  <li>o > ǫ (Ǫ)</li>
+  <li>s > s̲ (S̲)</li>
+  <li>t > t̲s̲ (T̲s̲ T̲S̲)</li>
+  <li>z > z̲ (Z̲)</li>
+</ul>
+
+<p>© 2015 FirstVoices, First Peoples' Cultural Foundation, SIL International</p>