+++ keyboard/redir.php
@@ -3,7 +3,7 @@
     header('Location: /keyboard');
     exit;
   }
-  
+
   $q = $_REQUEST['q'];
   if(preg_match('/[:\/]/', $q)) {
     header('HTTP/1.0 400 Invalid parameter');
@@ -22,16 +22,19 @@
       header('Location: /keyboard');
       exit;
     }
+    // We need to sort the returned array because it will be sorted by string
+    // which makes '10.0' < '9.0'.
+    usort($versions, 'version_compare');
     $v = array_pop($versions);
   }
-  
+
   $filename = "$v/$q.php";
   if(!file_exists($filename)) {
     //Debug purposes: echo "Failed to find target filename\n";
     header('Location: /keyboard');
     exit;
   }
-  
+
   //Debug purposes: echo "Location: /keyboard/$filename";
   header("Location: /keyboard/$filename");
 ?>